do not clean model dir by default

+++ models/DNN.py
@@ -20,7 +20,7 @@ DUMPDIR = "dump"
 OUTDIR = os.path.join("trained_models", "DNN")
 
 
-def model(df):
+def model(df, force_clean=False):
     user_vocabulary, assignee_vocabulary, most_active_vocabulary = vocabularies(df)
     os.makedirs(DUMPDIR, exist_ok=True)
 
@@ -36,7 +36,8 @@ def model(df):
     evaldf = df[~msk]
 
     # Run the model
-    shutil.rmtree(OUTDIR, ignore_errors=True)
+    if force_clean:
+        shutil.rmtree(OUTDIR, ignore_errors=True)
     train_and_evaluate(OUTDIR, 1000, most_active_vocabulary, user_vocabulary, traindf, evaldf)
